+++ OnionHarvester/OnionAddress.java
@@ -1,0 +1,7 @@
+package OnionHarvester;
+
+/**
+ * Created by Mir Saman on 01-Jan-18.
+ */
+public class OnionAddress {
+}
